Refactor(#39) : LevelManager 책임 분리
